refactor(editor): split GameDesignerWindow and reorganize Editor folder (#119)
- Folder restructuring: Builders/ (spawn/setup builders) + Windows/ (UI windows)
- GameDesignerWindow split into shell + TeamBuilderTab + LevelDesignerTab
- Removed #region directives throughout
- Tightened access modifiers (private by default)
- Applied DRY fixes to builder methods
- Dead code removal and cleanup
- Updated CLAUDE.md project structure
