+++ Assets/Scripts/Editor/Builders/BootstrapSceneBuilder.cs.meta
