+++ Assets/Scripts/Editor/Windows.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f6da04a5b7c189d72b3f4a6c8e1fa5b4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
